requirements.txt
asgiref==3.6.0
boto3==1.26.57
botocore==1.29.57
dj-database-url==1.2.0
Django==4.1.4
django-cors-headers==3.13.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
jmespath==1.0.1
Pillow==9.3.0
psycopg2==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
wincertstore==0.2
